irssi: embrace the Developer Certificate of Origin for contributions
The Linux kernel started to embrace a light weight development contribution
agreement process called the Developer Certificate of Origin (DCO) circa
2004. Attorneys all over the planet have conferred a lot of respect and
appreciation for this document and it provides a light weight alternative to
cumbersome and controversial Contribution License Agreements (CLAs).

Similar to streamlining FOSS licenses we also all stand to gain from
streamlining a light weight contribution agreement process for development
and in light of this the the Linux Foundation recently has made the DCO a
standalone project [0] so that any FOSS project under any FOSS license can
take advantage of the same gains. The motivation and intent is documented
on my blog post [1] and later covered extensively on lwn [2] after
James Bottomley's talk, and Bradley Kuhn went on extensively about it
as well [3].

Lets take advantage of the new shiny DCO and embrace it for contributions.

[0] http://developercertificate.org/
[1] http://www.do-not-panic.com/2014/02/developer-certificate-of-origin.html
[2] http://lwn.net/Articles/592503/
[3] http://ebb.org/bkuhn/blog/2014/06/09/do-not-need-cla.html

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>