irssi core: enable prompting for client certificate password
irssi supports letting you specify the client SSL password
using the ssl_pass primitive on the configuration file. Its
a terrible idea to be putting passwords on unencrypted files
though so some security rational folks won't ever want to
do this. If you currently don't supply one irssi goes on and
assumes you do not want to use it, and depending on the server
configuration it may or not fail on the SSL handshake and not
allow any connection to go through... Freenode lets you
authenticate your username with an SSL client certificate
so this is not desirable behavior.

This adds support to irrsi so that if an SSL client
certificate was supplied and if no password was set we prompt
the user for one.

Upon disconnects we won't have to re-enter the password but
just in case for whatever reason SSL needs it we redraw the
screen to the console screen should the user have to enter
a password again. Users that do not have a password on the
SSL client certificate can just either enter an empty
password upon the prompt or set an empty password on the
configuration file. Its a bad idea to be using passwordless
client certificate files anyway, so let them do a bit more
work.

We only prompt for a password if and only if the client
certificate was set, if you have different client certificate
files or use the client certificate for different connections
you'll be asked for a password for each server SSL connection.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>