Strip trailing whitespace from cmd args to avoid poor autocompletion UI
On my Android phone the keyboard will "helpfully" always add a trailing
space when I select one of the autocomplete suggestions anticipating that
I will be adding another word. This seems a sensible default option in a
phone environment.

However, a command such as "/win 4 " is not accepted by irssi, thus when
using the auto-completion I have to manually remove this trailing space if
I want it to work. This patch therefore removes the trailing whitespaces
from command arguments, improving the mobile user experience.

See also:

 * http://bugs.irssi.org/index.php?do=details&task_id=727
 * http://bugs.irssi.org/index.php?do=details&task_id=769

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>