Move new stuff at the end of IRC_SERVER_REC to attempt to unbreak ABI
This normally wouldn't matter at all but I just spent a couple of hours
figuring out a bug that turned out to be Server->isupport() from a
0.8.17 ABI Irc.so reading from the splits hash table instead of the
isupport hash table, and both happen to be 32 bytes away from each
other, which is the same size of the newly added struct items. So it
failed in a very elegant and subtle way (it would probably crash and
burn in 32bit architectures, though)

This commit is just an attempt to get something useful out of that time.

It also moves the cap_complete flag to the bitfield above so it can be
properly packed (taking one bit instead of 4 bytes like it did in the
other position). That's 4 bytes less per network! 60 bytes total for me!
Amazing.