Add support for SASL mechanisms handled by scripts
This implements the suggested API from #323 with a caveat: currently the
SASL mechanisms are tied to IRC and needs a bit of extra work to make
them generic enough to move to the regular src/core instead of solely in
src/irc/core.

This greatly simplifies scripts for each mechanism since this handles
the splitting and reassembly of the 400 byte fragments going to and
coming from the server.

The ECDSA-NIST256P-CHALLENGE mechanism has been reimplemented using this
new API and found to work.