Send 'setup changed' event after we echo the new setting
If the setting was created by a Perl script, and an error occurs while
processing the 'setup changed' signal, the Perl script will unload and
destroy the SETTINGS_REC out from under us, causing a use-after-free
and likely a segfault.

This commit reorders the calls so that the 'setup changed' signal
happens last.  As an added bonus, any output generated by any module
or script in response to the new setting will show up *after* we
confirm the setting change, rather than before.