rephrase documentation to be abstract to any given network
IRSSI is an IRC client and should not advertise nor be beholden to any
particular IRC network. This patch is inspired by both of the following
commits, and aims to remove references to any particular network in the
documentation, in an effort to avoid potential conflict amongst team and
community members:

- a4486c236a3bf15192d0500b3a1892f7465826c7
- 1ba48840a112dfacf13cbbf6b77c1e3489fefcf8

Based on a (as of this writing) draft blog post [0], and the discussion
that has occurred regarding the blog post, it doesn't appear as if the
IRSSI development team is interested in being transparent about the
reasoning behind its aversion to removing the Freenode network,
specifically related to the comments about "potential legal
consequences". By removing references to any given network in the
documentation, this project moves forward without taking a political
stance, but in a way that avoids recommending any given network to
users.

[0]: https://github.com/irssi/irssi.github.io/pull/79