build: fix linking error on arm64 Darwin
Currently, it is not possible to build git master on Darwin, as seen in
several issues reported previously.

This commit fixes that issues, by initializing two variables to NULL.
I have no idea, why this solves the linking error, as to my knowledge,
global variables do not need to be initialized.